% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huckster.R
\name{huc_code}
\alias{huc_code}
\title{Convert Labels, Hydrologic Unit Codes, or HUC Lengths to Layer IDs}
\usage{
huc_code(x)
}
\arguments{
\item{x}{One or more: labels, codes, or code lengths to convert to \code{layerid}}
}
\value{
integer
}
\description{
This function takes heterogeneous input and calculates the \code{layerid} required for creating the REST API requests for geometry.
}
\examples{

huc_code(c("subregion", "watershed",
           "071000050101", "10-digit",
           8, 12, 16))

}
