% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huckster-package.R
\docType{package}
\name{huckster-package}
\alias{huckster}
\alias{huckster-package}
\title{huckster: Very Simple Tools for Hawking Hydrologic Unit Boundaries}
\description{
Provides tools for easily obtaining boundaries of hydrologic units and information based on Hydrologic Unit Codes ('HUC'). Hydrologic unit data are retrieved from the USGS NationalMap ArcGIS MapServer web services: <"https://hydro.nationalmap.gov/arcgis/rest/services/wbd/MapServer/>.
}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

}
\keyword{internal}
