% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huckster.R
\name{point_to_huc}
\alias{point_to_huc}
\title{Intersect hydrologic unit boundaries at the specified level with a SpatVector containing points or X,Y coordinates}
\usage{
point_to_huc(
  x,
  y = NULL,
  layer = "watershed",
  layerid = huc_code(layer),
  sr_in = 4326,
  sr_out = sr_in,
  base_url = huckster_usgs_hydro_wbd_url()
)
}
\arguments{
\item{x}{A \code{SpatVector} object or \code{numeric} vector.}

\item{y}{\code{NULL} (when x is a spatial object); otherwise a \code{numeric} vector equal in length to \code{x}}

\item{layer}{One or more: labels, codes, or code lengths to convert to \code{layerid}. Default \code{"watershed"}, which is equivalent to "10-digit", "10" or any 10-digit code.}

\item{layerid}{numeric. \code{1:8} where \code{2*layerid} is equal to the number of digits in the Hydrologic Unit Code (HUC). Default: \code{5} derived from \code{"watershed"} for "10-digit" hydrologic unit boundary. See details.}

\item{sr_in}{integer. Spatial Reference System of input (\code{x}, \code{y}) as specified with numeric code. Default: \code{4326} for \code{"EPSG:4326"}.}

\item{sr_out}{integer. Spatial Reference System of result as specified with numeric code. Default: \code{sr_in}; equivalent to input.}

\item{base_url}{Default: \code{"https://hydro.nationalmap.gov/arcgis/rest/services/wbd/MapServer/"}}
}
\value{
A \code{SpatVector} object derived from GeoJSON.
}
\description{
Hydrologic unit boundaries are retrieved from the USGS NationalMap ArcGIS MapServer web services: \code{"https://hydro.nationalmap.gov/arcgis/rest/services/wbd/MapServer/"}.
}
\details{
The levels of \code{layerid} correspond to the numeric codes for the following hydrologic units (HU):
\itemize{
\item WBDLine (0)
\item 2-digit HU (Region) (1)
\item 4-digit HU (Subregion) (2)
\item 6-digit HU (Basin) (3)
\item 8-digit HU (Subbasin) (4)
\item 10-digit HU (Watershed) (5)
\item 12-digit HU (Subwatershed) (6)
\item 14-digit HU (7)
\item 16-digit HU (8)
}
}
\examples{
\dontshow{if (!inherits(requireNamespace("terra", quietly = TRUE), 'try-error')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
\dontrun{
point_to_huc(-120, 36:39)
}
}
